feat: Search for video via Youtube API
When a query is given search_youtube will attempt
to find a video. In the likely scenario that video is not embedabble,
app_routes.py will work with the js on index.html to embed the thumbnail
and link to youtube instead of embedding the video itself.

The embed is still ugly if the video is unavailable for other reasons.